Linking the progress view to the database
Only the necessary has been added to the viewmodels, there is still a lot of things to add.
The pattern used to create the viewModel doesn't seem to be perfect as well, i'll check it out with someone else to think about a pattern where the entity object is not given to each constructor.